+++ Tasty Fridge!/Model/FoodCategory+CoreDataProperties copy.swift
@@ -17,6 +17,6 @@ extension FoodCategory {
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int32
 
 }
